Rejigger entire project configuration
Created a new Tictactoe.hs which contains the majority of the code that
was in Game.hs and is a module that exports to Game.hs. Then Game.hs
contains only a main function. Removed much of the unused testing code
to Tests.hs and have it actually run every time.