Finished tentative version of SuperTicTacToe
It seems like SuperTicTacToe is not recognizing victory conditions
correctly, based on one playthrough; however, tests seem to be showing
that the victory conditions are in fact working. Still have not
implemented AI. Also have not implemented rule that if a player lands on
a full mini-board, that he/she can move wherever he/she desires.