Moved more functions to Negamax
Moved away Tictactoe-specific negamax functions, namely the findBestMove
function to Negamax.hs and made it independent from Tictactoe so that it
could be reused elsewhere, namely in SuperTictactoe. It is still wrapped
up in another findBestMove in Tictactoe.hs to maintain comptability with
the rest of the file.

This also meant that the generateNegamaxTree function in Tictactoe.hs
was no longer necessary, as findBestMove was more opaque.