Added pure AI game and simplified some functions
AIGame.hs is a temporary file that is used for a pure AI vs AI match.
Until I get a hang of optparse-applicative, this will be useful for
benchmarking speed.

I have also modified playMove so that it will not always check
isValidMove (playMove is used by both the computer and the player and
the computer will never enter an invalid move). This gives a small
performance boost to the AI.

sameEntries has also been modified to something which theoretically is
faster, but which hasn't really seen any performance boosts in real
life.