Removed use of OverlappingInstances
The consensus of the Haskell community seems to be that
OverlappingInstances is an unsafe extension to use. As such I've manned
up and removed it. This means the nice syntactic sugar of using the
simple show command for showing a GameBoardUnit has been replaced with a
less generic showGameBoardUnit function.