Added commandline-options
Now all of the SuperTictactoe and Tictactoe stuff can be accessed via
the same commandline program. This was done using optparse-applicative
to parse the various flags that are required to get the various variants
in play. Along the way, minor modifications were made in what gets
printed to the screen by functions in Tictactoe.hs and SuperTictactoe.hs
to make things prettier.