Added new showGameBoard function and Wall misc
As a result of trying to make showGameBoard total, I have refactored it
so that it can now be used on boards of arbitrary size instead of just
3x3 boards.

There were also other miscellaneous changes to correct warnings showing
up after -Wall. The biggest of these were providing inefficient, but
correct fallbacks for functions like getDiagonals and getCols to make
them total.