Make sure that we handle the case of a full board
There is the possibility that a player gets sent to a mini-board which
is already full. Before this would just result in a runtime error. Now
this is considered an illegal move. If no move that does not result in
this is possible, then we end the game as a tie.